removed sourceTypeLink management layer
